+++ guillotina/tests/test_postgres.py
@@ -1,12 +1,13 @@
-from guillotina.behaviors import apply_markers
 from guillotina.content import Folder
-from guillotina.content import Item
 from guillotina.db.storages.pg import PostgresqlStorage
 from guillotina.db.transaction_manager import TransactionManager
 from guillotina.exceptions import ConflictError
+from guillotina.db.resolution import record_object_change
+from guillotina.behaviors.dublincore import IDublinCore
+from guillotina.interfaces import IResource
+from guillotina.tests.utils import create_content
 
 import pytest
-import uuid
 
 
 async def cleanup(aps):
@@ -29,24 +30,6 @@ async def get_aps():
     return aps
 
 
-def create_ob():
-    obj = Item()
-    obj.type_name = 'Item'
-    obj._p_oid = uuid.uuid4().hex
-    obj.__name__ = obj.id = 'foobar'
-    apply_markers(obj, None)
-    return obj
-
-
-def create_folder():
-    obj = Folder()
-    obj.type_name = 'Folder'
-    obj._p_oid = uuid.uuid4().hex
-    obj.__name__ = obj.id = 'foobar'
-    apply_markers(obj, None)
-    return obj
-
-
 async def test_read_obs(postgres, dummy_request):
     """Low level test checks that root is not there"""
     request = dummy_request  # noqa so magically get_current_request can find
@@ -56,7 +39,7 @@ async def test_read_obs(postgres, dummy_request):
     await tm.begin()
     txn = tm._txn
 
-    ob = create_ob()
+    ob = create_content()
     txn.register(ob)
 
     assert len(txn.modified) == 1
@@ -86,9 +69,9 @@ async def test_deleting_parent_deletes_children(postgres, dummy_request):
     await tm.begin()
     txn = tm._txn
 
-    folder = create_folder()
+    folder = create_content(Folder, 'Folder')
     txn.register(folder)
-    ob = create_ob()
+    ob = create_content()
     await folder.async_set('foobar', ob)
 
     assert len(txn.modified) == 2
@@ -130,7 +113,7 @@ async def test_create_blob(postgres, dummy_request):
     await tm.begin()
     txn = tm._txn
 
-    ob = create_ob()
+    ob = create_content()
     txn.register(ob)
 
     await txn.write_blob_chunk('X' * 32, ob._p_oid, 0, b'foobar')
@@ -161,7 +144,7 @@ async def test_delete_resource_deletes_blob(postgres, dummy_request):
     await tm.begin()
     txn = tm._txn
 
-    ob = create_ob()
+    ob = create_content()
     txn.register(ob)
 
     await txn.write_blob_chunk('X' * 32, ob._p_oid, 0, b'foobar')
@@ -198,7 +181,8 @@ async def test_should_raise_conflict_error(postgres, dummy_request):
     await tm1.begin()
     txn = tm1._txn
 
-    ob = create_ob()
+    ob = create_content()
+    ob.title = 'foobar'
     txn.register(ob)
 
     await tm1.commit()
@@ -212,6 +196,7 @@ async def test_should_raise_conflict_error(postgres, dummy_request):
     ob1 = await txn1.get(ob._p_oid)
     ob2 = await txn2.get(ob._p_oid)
     ob1.title = 'foobar1'
+    record_object_change(ob1, IResource['title'], 'foobar')
     ob2.title = 'foobar2'
 
     txn1.register(ob1)
@@ -226,6 +211,75 @@ async def test_should_raise_conflict_error(postgres, dummy_request):
     await cleanup(aps)
 
 
+async def test_should_not_raise_conflict_error_when_editing_diff_data(postgres, dummy_request):
+    request = dummy_request  # noqa so magically get_current_request can find
+
+    aps = await get_aps()
+    tm1 = TransactionManager(aps)
+    tm2 = TransactionManager(aps)
+
+    # create object first, commit it...
+    await tm1.begin()
+    txn = tm1._txn
+
+    ob = create_content()
+    ob.title = 'foobar'
+    ob.description = 'foobar'
+    txn.register(ob)
+
+    await tm1.commit()
+
+    # 1 started before 2
+    await tm1.begin()
+    await tm2.begin()
+    txn1 = tm1._txn
+    txn2 = tm2._txn
+
+    ob1 = await txn1.get(ob._p_oid)
+    ob2 = await txn2.get(ob._p_oid)
+    ob1.title = 'foobar1'
+    record_object_change(ob1, IResource['title'], 'foobar')
+    ob2.description = 'foobar2'
+    record_object_change(ob2, IDublinCore['description'], 'foobar')
+
+    txn1.register(ob1)
+    txn2.register(ob2)
+
+    # commit 2 before 1
+    await tm2.commit()
+    await tm1.commit()
+
+    # and new object should have data from each
+    await tm1.begin()
+    ob1 = await txn1.get(ob._p_oid)
+    assert ob1.title == 'foobar1'
+    assert ob2.description == 'foobar2'
+
+    await tm1.commit()
+
+    await aps.remove()
+    await cleanup(aps)
+
+
+async def test_should_reset_changes_after_commit(postgres, dummy_request):
+    request = dummy_request  # noqa so magically get_current_request can find
+
+    aps = await get_aps()
+    tm = TransactionManager(aps)
+    await tm.begin()
+    txn = tm._txn
+
+    ob = create_content()
+    ob.title = 'foobar'
+    record_object_change(ob, IResource['title'], 'foobar')
+    txn.register(ob)
+
+    assert len(txn.modified) == 1
+    assert len(ob.__changes__) == 1
+    await tm.commit()
+    assert len(ob.__changes__) == 0
+
+
 async def test_should_resolve_conflict_error(postgres, dummy_request):
     request = dummy_request  # noqa so magically get_current_request can find
 
@@ -237,8 +291,8 @@ async def test_should_resolve_conflict_error(postgres, dummy_request):
     await tm1.begin()
     txn = tm1._txn
 
-    ob1 = create_ob()
-    ob2 = create_ob()
+    ob1 = create_content()
+    ob2 = create_content()
     txn.register(ob1)
     txn.register(ob2)
 
@@ -275,7 +329,7 @@ async def test_count_total_objects(postgres, dummy_request):
     await tm1.begin()
     txn = tm1._txn
 
-    ob = create_ob()
+    ob = create_content()
     txn.register(ob)
 
     await tm1.commit()